feat : 단건 조회

+++ src/main/java/com/mink/freshexpress/category/dto/SimpleCategoryResponseDto.java
@@ -7,7 +7,7 @@ import lombok.Getter;
 public class SimpleCategoryResponseDto {
     private final Long id;
     private final String name;
-    private final Long depth;
+    private final int depth;
 
     public SimpleCategoryResponseDto(Category category) {
         this.name = category.getName();
